[CLEANUP] 코드 분리 작업 1차
- 진단/학습/노트 분리
- 기능별 분리
(현재 sidevar 오류로 해결 중)
