+++ package.json
@@ -52,7 +52,7 @@
     "react-day-picker": "9.8.0",
     "react-dom": "19.2.0",
     "react-hook-form": "^7.60.0",
-    "react-resizable-panels": "^2.1.7",
+    "react-resizable-panels": "^4.7.3",
     "recharts": "2.15.4",
     "sonner": "^1.7.4",
     "tailwind-merge": "^3.3.1",
